+++ src/__tests__/chat_stream_handlers.test.ts
@@ -9,7 +9,7 @@ import {
 import fs from "node:fs";
 import git from "isomorphic-git";
 import { db } from "../db";
-import { cleanFullResponse } from "@/ipc/handlers/cleanFullResponse";
+import { cleanFullResponse } from "@/ipc/utils/cleanFullResponse";
 
 // Mock fs with default export
 vi.mock("node:fs", async () => {
