+++ src/ipc/utils/cleanFullResponse.ts
